+++ src/character/character_store.cpp
@@ -22,6 +22,8 @@
 #include "../events/event_queue.h"
 #include "../env/env.h"
 #include "../dungeon/pathfinder.h"
+#include "../parser/parser.h"
+#include "../parser/monster_description.h"
 
 int character_store::CHARACTER_COUNT = 0;
 int character_store::_characters_size = 0;
@@ -33,9 +35,9 @@ void character_store::print_char(character* npc) {
     point* npc_pos;
     point* npc_dest;
     uint8_t  npc_attrs;
-    npc_pos   = npc->_position;
-    npc_dest  = npc->_destination;
-    npc_attrs = npc->_attrs;
+    npc_pos   = npc->position;
+    npc_dest  = npc->destination;
+    npc_attrs = npc->attrs;
     if(env_constants::DEBUG_MODE && !env_constants::NCURSES_MODE) {
         printf("NPC: spawn: (%2d, %2d) ",
                npc_pos->x,
@@ -65,27 +67,20 @@ void character_store::print_char(character* npc) {
 void character_store::setup() {
     int nummon = env_constants::NUM_MONSTERS;
     int i;
-    dungeon* d = dungeon::get_dungeon();
     _characters_size = nummon; // num of monsters
     _characters = (character**) calloc(_characters_size, sizeof(*_characters));
     _characters_count = _characters_size;
     CHARACTER_COUNT = _characters_count;
     _alive_characters = (character_id_t*) calloc(_characters_count, sizeof(*_alive_characters));
-    character* pc = character::get_pc();
-    point* pc_pos = pc->_position;
     if(env_constants::PC_AI_MODE) {
         ai::setup_pc_movement();
     } else {
         pc_control::setup_control_movement();
     }
     for(i = 0; i < _characters_size; i++) {
-        point* spawn = new point(0, 0);
-        do {
-            dungeon::rand_point(d, spawn);
-        } while(spawn->distance_to(pc_pos) <= 3); // have a radius of 3 blocks
         
-        character* npc = new character(NPC, spawn);
-        npc->_id = i+1;
+        character* npc = gen_npc(env_constants::USE_MPARSE);
+        npc->id = i+1;
         
         setup_npc(npc);
         
@@ -93,7 +88,6 @@ void character_store::setup() {
         _characters[i] = npc;
         _alive_characters[i] = i+1;
         
-        free(spawn);
         print_char(npc);
     }
 }
@@ -118,11 +112,11 @@ int character_store::contains_npc(point* p) {
     point* npc_pos;
     point* pc_pos;
     uint8_t  npc_is_dead;
-    pc_pos = character::get_pc()->_position;
+    pc_pos = character::get_pc()->position;
     // check npcs first, then player
     for(i = 0; i < _characters_count; i++) {
-        npc_pos = _characters[i]->_position;
-        npc_is_dead = _characters[i]->_is_dead;
+        npc_pos = _characters[i]->position;
+        npc_is_dead = _characters[i]->is_dead;
         if(p->distance_to(npc_pos) == 0 && !npc_is_dead) {
             return i+1;
         }
@@ -133,24 +127,28 @@ int character_store::contains_npc(point* p) {
     return -1;
 }
 
-char character_store::get_char_for_npc_at_index(int i) {
+char character_store::get_char_for_npc_at_index(int i, int mode) {
     if(i == 0) {
-        return character::get_pc()->char_for_npc_type();
+        return character::get_pc()->get_print_symb(mode);
     }
-    return _characters[i-1]->char_for_npc_type();
+    return _characters[i-1]->get_print_symb(mode);
 }
 
 character** character_store::get_characters() {
     return _characters;
 }
 
+character_id_t* character_store::get_alive_characters() {
+    return _alive_characters;
+}
+
 void character_store::setup_npc(character* npc) {
     point* pc_pos;
     point* npc_pos;
     uint8_t npc_attrs;
-    pc_pos = character::get_pc()->_position;
-    npc_pos = npc->_position;
-    npc_attrs = npc->_attrs;
+    pc_pos = character::get_pc()->position;
+    npc_pos = npc->position;
+    npc_attrs = npc->attrs;
     // telepathic npcs get the pc position
     if(npc_attrs & TELEP_VAL) {
         npc->set_destination(pc_pos);
@@ -169,15 +167,15 @@ int character_store::is_finished() {
     point* pc_pos;
     point* npc_pos;
     uint8_t  npc_is_dead;
-    pc_pos = character::get_pc()->_position;
+    pc_pos = character::get_pc()->position;
     // only 1 character left (pc) so pc has won
     if(_characters_count == 0) {
         return 2;
     }
     // check all npcs that ARE NOT the pc for collision
     for(i = 0; i < _characters_size; i++) {
-        npc_pos = _characters[i]->_position;
-        npc_is_dead = _characters[i]->_is_dead;
+        npc_pos = _characters[i]->position;
+        npc_is_dead = _characters[i]->is_dead;
         if(pc_pos->distance_to(npc_pos) == 0 && !npc_is_dead) {
             return 1;
         }
@@ -194,7 +192,7 @@ void character_store::npc_cleanup() {
     for(i = 0; i < _characters_count; i++) {
         npc_is_dead = 0;
         character* npc = npc_for_id(_alive_characters[i]);
-        npc_is_dead = npc->_is_dead;
+        npc_is_dead = npc->is_dead;
         if(npc_is_dead) {
             logger::d("npc %d is dead, shifting over", _alive_characters[i]);
             // shift over other npcs
@@ -204,6 +202,7 @@ void character_store::npc_cleanup() {
             _characters_count--;
         }
     }
+    CHARACTER_COUNT = _characters_count;
     logger::d("NPC cleanup: %d ~> %d", old_count, _characters_count);
 }
 
@@ -222,18 +221,18 @@ void character_store::start_monster_list() {
         return; // this can only be shown when using pc control, which requires NCURSES_MODE to be active.
     }
     
-    pc_pos = character::get_pc()->_position;
+    pc_pos = character::get_pc()->position;
     monster_list = (char**) calloc(_characters_count, sizeof(*monster_list));
     for(i = 0; i < _characters_count; i++) {
         character* npc = npc_for_id(_alive_characters[i]);
         if(npc != NULL) {
-            npc_pos = npc->_position;
+            npc_pos = npc->position;
             monster_list[i] = (char*) calloc(30, sizeof(**monster_list));
             xdiff = pc_pos->x - npc_pos->x;
             ydiff = pc_pos->y - npc_pos->y;
             sprintf(monster_list[i], "%2d. %c, %2d %s and %2d %s",
                     i+1,
-                    npc->char_for_npc_type(),
+                    npc->symb,
                     abs(ydiff),
                     ydiff > 0 ? "north" : "south",
                     abs(xdiff),
@@ -284,10 +283,31 @@ character* character_store::npc_for_id(character_id_t id) {
     int i;
     uint8_t npc_id;
     for(i = 0; i < _characters_size; i++) {
-        npc_id = _characters[i]->_id;
+        npc_id = _characters[i]->id;
         if(id == npc_id) {
             return _characters[i];
         }
     }
     return NULL;
+}
+
+character* character_store::gen_npc(bool use_descriptor) {
+    character* pc = character::get_pc();
+    point* pc_pos = pc->position;
+    dungeon* d = dungeon::get_dungeon();
+    point spawn(0, 0);
+    do {
+        dungeon::rand_point(d, &spawn);
+    } while(spawn.distance_to(pc_pos) <= 3); // have a radius of 3 blocks
+    
+    monster_description** monster_list = parser::get_parser()->monster_list;
+    
+    if(!use_descriptor || monster_list == NULL) {
+        return new character(NPC, &spawn);
+    }
+    
+    int rand_idx = rand() % parser::get_parser()->monster_len;
+    monster_description* descriptor = monster_list[rand_idx];
+    
+    return new character(NPC, &spawn, descriptor);
 }