+++ src/character/character_store.h
@@ -20,8 +20,8 @@ private:
     static int _characters_count;
     static void print_char(character* npc);
     static void setup_npc(character* npc);
-    static character* npc_for_id(character_id_t id);
     
+    static character* gen_npc(bool use_descriptor);
     
 public:
     static int CHARACTER_COUNT;
@@ -29,8 +29,10 @@ public:
     static void setup();
     static void teardown();
     static int  contains_npc(point* p);
-    static char get_char_for_npc_at_index(int i);
+    static char get_char_for_npc_at_index(int i, int mode);
     static character** get_characters();
+    static character_id_t* get_alive_characters();
+    static character* npc_for_id(character_id_t id);
     static int  is_finished();
     static void npc_cleanup();
     static void start_monster_list();