+++ src/character/pc_control.cpp
@@ -13,6 +13,7 @@
 #include "../events/event_queue.h"
 #include "../logger/logger.h"
 #include "../dungeon/dungeon.h"
+#include "../items/item_store.h"
 
 void pc_control::setup_control_movement() {
     event_queue::add_event(character::get_pc());
@@ -25,7 +26,7 @@ void pc_control::handle_control_move() {
     point* pc_pos;
     point* npc_pos;
     pc = character::get_pc();
-    pc_pos = pc->_position;
+    pc_pos = pc->position;
     pc_move_t move = mv_NONE;
     point* dest = new point(pc_pos);
     int i;
@@ -169,17 +170,23 @@ void pc_control::handle_control_move() {
         logger::i("Moving to point (%2d, %2d)", dest->x, dest->y);
         // move pc
         pc->set_position(dest);
-        pc_pos = pc->_position;
+        pc_pos = pc->position;
         
         // check for collision
         for(i = 0; i < character_store::CHARACTER_COUNT; i++) {
-            npc_pos = characters[i]->_position;
+            npc_pos = characters[i]->position;
             if(pc_pos->distance_to(npc_pos) == 0) {
-                characters[i]->_is_dead = 1;
+                characters[i]->is_dead = 1;
             }
+            characters[i]->is_seen = (pc_pos->distance_to(npc_pos) <= 3);
         }
     } else if(move == mv_RS) {
         logger::i("Resting 1 turn");
+    }
+    
+    // check items
+    if(item_store::contains_item(pc->position)) {
+        item_store::pickup_item(pc->position);
     }
     
     delete dest;