+++ src/env/env.cpp
@@ -12,6 +12,7 @@
 #include <string.h>
 #include <getopt.h>
 #include <errno.h>
+#include <time.h>
 #include <sys/stat.h>
 #include <ncurses.h>
 
@@ -20,8 +21,14 @@
 #include "../character/character_store.h"
 #include "../dungeon/dungeon.h"
 #include "../events/event_queue.h"
+#include "../parser/parser.h"
+#include "../items/item_store.h"
 
-int env_constants::PARSE_MODE = 0;
+int env_constants::USE_IPARSE = 1;
+int env_constants::USE_MPARSE = 1;
+int env_constants::USE_OBJ_DESC = 1; // use parsed objects by default
+int env_constants::USE_MON_DESC = 1; // use parsed monsters by default
+int env_constants::PARSE_ONLY_MODE = 0;
 int env_constants::DEBUG_MODE = 0;
 int env_constants::NCURSES_MODE = 0; // pc control enabled by default (so ncurses is enabled implicitly)
 int env_constants::PC_AI_MODE = 0;
@@ -44,6 +51,7 @@ static const char* help_text = "Usage: hexterm_havoc [options]\n\n"
                                "-l<name>, --load=<name> | Load dungeon with name <name> (in save directory).\n"
                                "-m<val> , --nummon=<val>| Set the number of monsters in the dungeon\n"
                                "-n      , --ncurses     | Use Ncurses to render game\n"
+                               "-p      , --parse-only  | Parse Objects and Monsters, then print them out\n"
                                "-s<name>, --save=<name> | Save the dungeon after loading/generating it with\n"
                                "                        |   name <name> (in save directory).\n"
                                "-x<val> , --xpos <val>  | Start the player at a specified x coord\n"
@@ -51,6 +59,8 @@ static const char* help_text = "Usage: hexterm_havoc [options]\n\n"
 
 void env::setup_environment() {
     logger::i("%%%% SETTING ENVIRONMENT %%%%");
+    srand((unsigned)time(NULL));
+    
     char* env;
     if((env = getenv("ENV"))) {
         if(!strcmp(env, "DEBUG")) {
@@ -66,7 +76,7 @@ void env::setup_environment() {
         env_constants::NCURSES_MODE = 1; // pc control means we must use NCURSES_MODE
     }
     
-    if(env_constants::PARSE_MODE) {
+    if(env_constants::PARSE_ONLY_MODE) {
         env_constants::NCURSES_MODE = 0;
     }
     
@@ -80,6 +90,15 @@ void env::setup_environment() {
         noecho();
         cbreak();
         set_escdelay(50);
+        start_color();
+        init_pair(COLOR_BLACK,   COLOR_WHITE,   COLOR_BLACK);
+        init_pair(COLOR_RED,     COLOR_RED,     COLOR_BLACK);
+        init_pair(COLOR_GREEN,   COLOR_GREEN,   COLOR_BLACK);
+        init_pair(COLOR_YELLOW,  COLOR_YELLOW,  COLOR_BLACK);
+        init_pair(COLOR_BLUE,    COLOR_BLUE,    COLOR_BLACK);
+        init_pair(COLOR_MAGENTA, COLOR_MAGENTA, COLOR_BLACK);
+        init_pair(COLOR_CYAN,    COLOR_CYAN,    COLOR_BLACK);
+        init_pair(COLOR_WHITE,   COLOR_BLACK,   COLOR_WHITE);
         keypad(stdscr, TRUE);
         mvprintw(0, 0, "DEBUG MODE");
         refresh();
@@ -110,19 +129,20 @@ void env::parse_args(int argc, char** argv) {
     while(1) {
         // Setup options
         static struct option long_options[] = {
-            {"ai",      no_argument,       0, 'a'},
-            {"load",    optional_argument, 0, 'l'},
-            {"save",    optional_argument, 0, 's'},
-            {"nummon",  required_argument, 0, 'm'},
-            {"ncurses", no_argument,       0, 'n'},
-            {"help",    no_argument,       0, 'h'},
-            {"xpos",    required_argument, 0, 'x'},
-            {"ypos",    required_argument, 0, 'y'},
+            {"ai",         no_argument,       0, 'a'},
+            {"parse-only", no_argument,       0, 'p'},
+            {"load",       optional_argument, 0, 'l'},
+            {"save",       optional_argument, 0, 's'},
+            {"nummon",     required_argument, 0, 'm'},
+            {"ncurses",    no_argument,       0, 'n'},
+            {"help",       no_argument,       0, 'h'},
+            {"xpos",       required_argument, 0, 'x'},
+            {"ypos",       required_argument, 0, 'y'},
             {0, 0, 0, 0}
         };
         
         int option_index = 0;
-        flag = getopt_long(argc, argv, "ahnm:x:y:s::l::", long_options, &option_index);
+        flag = getopt_long(argc, argv, "pahnm:x:y:s::l::", long_options, &option_index);
         
         if(flag == -1) {
             break;
@@ -176,6 +196,10 @@ void env::parse_args(int argc, char** argv) {
                 
             case 'n':
                 env_constants::NCURSES_MODE = 1;
+                break;
+                
+            case 'p':
+                env_constants::PARSE_ONLY_MODE = 1;
                 break;
                 
             case 'l':
@@ -247,8 +271,10 @@ void env::cleanup() {
     }
     
     character_store::teardown();
+    item_store::teardown();
     event_queue::teardown();
     dungeon::teardown();
+    parser::destroy_parser();
     
     if(env_constants::NCURSES_MODE) {
         endwin();