+++ src/parser/dice.cpp
@@ -8,6 +8,7 @@
 
 #include <string>
 #include <stdio.h>
+#include <stdlib.h>
 
 #include "dice.h"
 #include "../logger/logger.h"
@@ -18,6 +19,21 @@ dice::dice(std::string str) {
     sides = 0;
     char buf[70]; // large enough to capture all beginning space;
     sscanf(str.c_str(), "%s%d+%dd%d", buf, &base, &num_dice, &sides);
+}
+
+dice::dice(dice* dice) {
+    num_dice = dice->num_dice;
+    base     = dice->base;
+    sides    = dice->sides;
+}
+
+int dice::roll() {
+    int roll = base;
+    int i;
+    for(i = 0; i < num_dice; i++) {
+        roll += (rand() % sides) + 1;
+    }
+    return roll;
 }
 
 std::string dice::to_string() {